requirements.txt
python-docx==0.8.11
Pillow==8.1.2
opencv-python==4.9.0.80
tqdm==4.66.1

#pytesseract==0.3.7
